+++ hatch/utils.py
@@ -4,6 +4,7 @@ import platform
 import re
 import shutil
 import subprocess
+import time
 from base64 import urlsafe_b64encode
 from datetime import datetime
 from contextlib import contextmanager
@@ -68,6 +69,10 @@ def get_admin_command():  # no cov
         return ['sudo', '-H'] + (['--user={}'.format(admin)] if admin else [])
 
 
+def is_project(d=None):
+    return os.path.isfile(os.path.join(d or os.getcwd(), 'setup.py'))
+
+
 def is_os_64bit():  # no cov
     # https://stackoverflow.com/a/12578715/5854007
     return platform.machine().endswith('64')
@@ -79,6 +84,17 @@ def conda_available():  # no cov
     except subprocess.CalledProcessError:
         return False
     return True
+
+
+def wait_until(f, *args):  # no cov
+    # https://github.com/kennethreitz/pipenv/pull/403
+    end_time = time.time() + 300
+    while time.time() < end_time:
+        if f(*args):
+            time.sleep(0.5)
+            return True
+        time.sleep(0.2)
+    raise AssertionError('timeout')
 
 
 def get_requirements_file(d, dev=False):