+++ tests/commands/test_clean.py
@@ -43,6 +43,50 @@ def test_cwd():
         assert_files_exist(files)
 
 
+def test_project_ignore_venv():
+    with temp_chdir() as d:
+        runner = CliRunner()
+        runner.invoke(hatch, ['init', 'ok', '--basic', '-ne'])
+        files = find_all_files(d)
+
+        test_file1 = os.path.join(d, 'test.pyc')
+        test_file2 = os.path.join(d, 'venv', 'test.pyc')
+        create_file(test_file1)
+        create_file(test_file2)
+        assert os.path.exists(test_file1)
+        assert os.path.exists(test_file2)
+
+        result = runner.invoke(hatch, ['clean'])
+
+        assert result.exit_code == 0
+        assert 'Cleaned!' in result.output
+        assert not os.path.exists(test_file1)
+        assert os.path.exists(test_file2)
+        assert_files_exist(files)
+
+
+def test_project_venv_no_detect():
+    with temp_chdir() as d:
+        runner = CliRunner()
+        runner.invoke(hatch, ['init', 'ok', '--basic', '-ne'])
+        files = find_all_files(d)
+
+        test_file1 = os.path.join(d, 'test.pyc')
+        test_file2 = os.path.join(d, 'venv', 'test.pyc')
+        create_file(test_file1)
+        create_file(test_file2)
+        assert os.path.exists(test_file1)
+        assert os.path.exists(test_file2)
+
+        result = runner.invoke(hatch, ['clean', '-nd'])
+
+        assert result.exit_code == 0
+        assert 'Cleaned!' in result.output
+        assert not os.path.exists(test_file1)
+        assert not os.path.exists(test_file2)
+        assert_files_exist(files)
+
+
 def test_cwd_compiled_only():
     with temp_chdir() as d:
         runner = CliRunner()
@@ -66,6 +110,66 @@ def test_cwd_compiled_only():
         assert not os.path.exists(test_file1)
         assert not os.path.exists(test_file2)
         assert not os.path.exists(test_file3)
+        assert_files_exist(files)
+
+
+def test_compiled_only_project_ignore_venv():
+    with temp_chdir() as d:
+        runner = CliRunner()
+        runner.invoke(hatch, ['init', 'ok', '--basic', '-ne'])
+        files = find_all_files(d)
+
+        test_file1 = os.path.join(d, 'test.pyc')
+        test_file2 = os.path.join(d, 'ok', 'test.pyc')
+        test_file3 = os.path.join(d, 'ok', 'deeper', 'test.pyc')
+        test_file4 = os.path.join(d, 'venv', 'test.pyc')
+        create_file(test_file1)
+        create_file(test_file2)
+        create_file(test_file3)
+        create_file(test_file4)
+        assert os.path.exists(test_file1)
+        assert os.path.exists(test_file2)
+        assert os.path.exists(test_file3)
+        assert os.path.exists(test_file4)
+
+        result = runner.invoke(hatch, ['clean', '-c'])
+
+        assert result.exit_code == 0
+        assert 'Cleaned!' in result.output
+        assert not os.path.exists(test_file1)
+        assert not os.path.exists(test_file2)
+        assert not os.path.exists(test_file3)
+        assert os.path.exists(test_file4)
+        assert_files_exist(files)
+
+
+def test_compiled_only_project_venv_no_detect():
+    with temp_chdir() as d:
+        runner = CliRunner()
+        runner.invoke(hatch, ['init', 'ok', '--basic', '-ne'])
+        files = find_all_files(d)
+
+        test_file1 = os.path.join(d, 'test.pyc')
+        test_file2 = os.path.join(d, 'ok', 'test.pyc')
+        test_file3 = os.path.join(d, 'ok', 'deeper', 'test.pyc')
+        test_file4 = os.path.join(d, 'venv', 'test.pyc')
+        create_file(test_file1)
+        create_file(test_file2)
+        create_file(test_file3)
+        create_file(test_file4)
+        assert os.path.exists(test_file1)
+        assert os.path.exists(test_file2)
+        assert os.path.exists(test_file3)
+        assert os.path.exists(test_file4)
+
+        result = runner.invoke(hatch, ['clean', '-c', '-nd'])
+
+        assert result.exit_code == 0
+        assert 'Cleaned!' in result.output
+        assert not os.path.exists(test_file1)
+        assert not os.path.exists(test_file2)
+        assert not os.path.exists(test_file3)
+        assert not os.path.exists(test_file4)
         assert_files_exist(files)