+++ tests/utils.py
@@ -2,10 +2,15 @@ import os
 import re
 import socket
 import time
+import traceback
 
 import pytest
 
 from hatch.env import get_package_version
+
+
+def print_traceback(exc_info):  # no cov
+    traceback.print_exception(*exc_info)
 
 
 def wait_for_os(s=None):
@@ -14,7 +19,7 @@ def wait_for_os(s=None):
 
 def wait_until(f, *args):  # no cov
     # https://github.com/kennethreitz/pipenv/pull/403
-    end_time = time.time() + 120
+    end_time = time.time() + 300
     while time.time() < end_time:
         if f(*args):
             time.sleep(0.5)