+++ HISTORY.rst
@@ -6,6 +6,13 @@ Important changes are emphasized.
 master
 ^^^^^^
 
+0.18.0
+^^^^^^
+
+- ``release`` now allows the use of custom repositories!
+- **Fix:** ``clean``\ ing now correctly ignores a project's dedicated virtual
+  env. This behavior can be disabled with the new ``-nd/--no-detect`` flag.
+
 0.17.1
 ^^^^^^